refactor: change group & namespace
io.github.jindong -> io.github.compose.jindong
