+++ jindong/src/commonMain/kotlin/io/github/compose/jindong/node/SequenceNode.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.jindong.node
+package io.github.compose.jindong.node
 
 /**
  * A container node that executes its children sequentially.
