+++ jindong/src/commonTest/kotlin/io/github/compose/jindong/node/HapticEventNodeTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.jindong.node
+package io.github.compose.jindong.node
 
 import io.kotest.assertions.assertSoftly
 import io.kotest.core.spec.style.FunSpec
